Send data with 16bit length, removed unfixed-length data

+++ Manchester.h
@@ -141,14 +141,14 @@ class Manchester
     void setup(uint8_t Tpin, uint8_t Rpin, uint8_t SF = MAN_1200); //set up receiver
     
     void transmit(uint16_t data); //transmit 16 bits of data
-    void transmitArray(uint8_t numBytes, uint8_t *data); // transmit array of bytes
+    void transmitArray(uint16_t numBytes, uint8_t *data); // transmit array of bytes
     
     uint8_t decodeMessage(uint16_t m, uint8_t &id, uint8_t &data); //decode 8 bit payload and 4 bit ID from the message, return 1 of checksum is correct, otherwise 0
     uint16_t encodeMessage(uint8_t id, uint8_t data); //encode 8 bit payload, 4 bit ID and 4 bit checksum into 16 bit
     
     //wrappers for global functions
     void beginReceive(void);
-    void beginReceiveArray(uint8_t maxBytes, uint8_t *data);
+    void beginReceiveArray(uint16_t maxBytes, uint8_t *data);
     uint8_t receiveComplete(void);
     uint16_t getMessage(void);
     void stopReceive(void);
@@ -177,7 +177,7 @@ extern "C"
     extern void MANRX_BeginReceive(void);
     
     // begin receiving a byte array
-    extern void MANRX_BeginReceiveBytes(uint8_t maxBytes, uint8_t *data);
+    extern void MANRX_BeginReceiveBytes(uint16_t maxBytes, uint8_t *data);
     
     // true if a complete message is ready
     extern uint8_t MANRX_ReceiveComplete(void);
